requirements.txt
# Chatterbox TTS Core Package with multilingual support (23 languages)
# Official ResembleAI version v0.1.4 - supports all 23 languages including Chinese
chatterbox-tts==0.1.4

# Required for Chatterbox TTS watermarker functionality
resemble-perth

# PyTorch with CPU support (ensure CPU compatibility)
torch>=2.0.0,<2.7.0
torchaudio>=2.0.0,<2.7.0

# FastAPI and ASGI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Required for FastAPI file upload support
python-multipart>=0.0.6

# Server-Sent Events support for real-time progress updates
sse-starlette>=3.0.2

# Environment variables
python-dotenv>=1.0.0

# Memory monitoring
psutil>=5.9.0

# Audio processing for long text concatenation
pydub>=0.25.1

# Testing Dependencies
requests>=2.28.0 